added top menu; added preferences for disabling the top menu, side menu and collapsing the side menu; added new settings page

+++ src/Commands/phantom.php
@@ -82,7 +82,8 @@ class phantom extends Command {
             shell_exec("npm install");
             shell_exec("npm run dev");
         }
-        unlink('phantom.setup.ready');
+        if (!$this->option('update'))
+            unlink('phantom.setup.ready');
 
     }
 }
